adds context to apolloServer, edits file structure
