+++ src/resolvers/index.js
@@ -1,0 +1,7 @@
+const Query = require('./query');
+const Mutation = require('./mutation');
+
+module.exports = {
+    Query,
+    Mutation
+}
