+++ src/resolvers/mutation.js
@@ -1,0 +1,8 @@
+module.exports = {
+    newNote: async (parent, args, { models }) => {
+        return await models.Note.create({
+            content: args.content,
+            author: 'Leslie Roriguez'
+        });
+    }
+}
