+++ src/resolvers/query.js
@@ -1,0 +1,8 @@
+module.exports = {
+    notes: async (parent, args, { models }) => {
+        return await models.Note.find();
+    },
+    note: async (parent, args, { models }) => {
+        return await models.Note.findById(args.id);
+    },
+}
