+++ src/schema.js
@@ -1,0 +1,21 @@
+const { gql } = require('apollo-server-express');
+
+const typeDefs = gql`
+    type Query {
+        hello: String
+        notes: [Note!]!
+        note(id: ID!): Note
+    }
+
+    type Note {
+        id: ID!
+        content: String!
+        author: String!
+    }
+
+    type Mutation {
+        newNote(content: String): Note!
+    }
+`;
+
+module.exports = typeDefs;
